services/backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
tortoise-orm[asyncpg]==0.20.0
python-dotenv==0.19.0
aiohttp==3.8.5 --pre
certifi==2021.10.8
redis>=5.0.0
hiredis>=2.0.0
aiokafka==0.7.2
kafka-python==2.0.2
confluent-kafka==2.3.0
aiogram==2.25.1 